+++ static/styles.css
@@ -1,0 +1,18 @@
+#guesses {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    height: 500px;
+    width: 100%;
+    overflow-y: auto;
+}
+
+#upload {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    height: 300px;
+    width: 50%;
+    margin: auto;
+    text-align: center;
+}
