+++ trainmodel.py
@@ -22,7 +22,7 @@ def build_model():
         input_shape =(3,IMG_WIDTH, IMG_HEIGHT)
     else:
         input_shape =(IMG_WIDTH, IMG_HEIGHT, 3)
-    #sequential data is the CNN model architecture that improves sample diversity, leading to improved test performance, especially in deep neural network
+    #sequential data augmentation substantially improves sample diversity, leading to improved test performance, especially in deep neural network
     model = Sequential()
     model.add(Conv2D(32, (3,3), input_shape=input_shape))
     model.add(Activation('relu'))
@@ -45,7 +45,9 @@ def build_model():
     
     model.compile(loss='binary_crossentropy',
                 optimizer ='rmsprop',
+                run_eagerly=True,
                 metrics=['accuracy'])
+    model.run_eagerly = True
     return model
